ancient_artisans/utils/ai_helpers.py
import google.generativeai as genai
import os
from dotenv import load_dotenv

load_dotenv()

# Configure Gemini API
GEMINI_API_KEY = os.getenv('GEMINI_API_KEY')
if GEMINI_API_KEY:
    genai.configure(api_key=GEMINI_API_KEY)

def analyze_product_image(image_url):
    """
    Analyze product image using Gemini AI and suggest a price
    In a real implementation, this would use the Gemini API
    """
    # For demo purposes, we'll return a simulated response
    # In production, you would use:
    # model = genai.GenerativeModel('gemini-pro-vision')
    # response = model.generate_content([prompt, image])
    
    return {
        "analysis": "This appears to be a handcrafted pottery vase with intricate designs. The craftsmanship is excellent with attention to detail. Based on similar products in the market, the materials used, and the uniqueness of the design, this is a fair market price.",
        "suggested_price": 79.99,
        "confidence": 0.87
    }

def generate_tutorial_description(product):
    """
    Generate tutorial description using Gemini AI
    In a real implementation, this would use the Gemini API
    """
    # For demo purposes, we'll return a simulated response
    # In production, you would use:
    # model = genai.GenerativeModel('gemini-pro')
    # response = model.generate_content(prompt)
    
    return {
        "title": f"How to make {product['name']}",
        "description": f"A step-by-step guide to creating {product['name']} using traditional techniques.",
        "steps": [
            {
                "step_number": 1,
                "title": "Prepare materials",
                "description": f"Gather all necessary materials including {product.get('materials', 'the required materials')}",
                "tips": ["Ensure materials are of high quality", "Organize your workspace before starting"]
            },
            {
                "step_number": 2,
                "title": "Basic shaping",
                "description": "Begin shaping the primary form of the product",
                "tips": ["Work slowly and carefully", "Refer to reference images if available"]
            },
            {
                "step_number": 3,
                "title": "Adding details",
                "description": "Add intricate details and patterns to enhance the design",
                "tips": ["Use specialized tools for fine details", "Take breaks to maintain precision"]
            },
            {
                "step_number": 4,
                "title": "Finishing touches",
                "description": "Apply final finishes and ensure quality control",
                "tips": ["Inspect from all angles", "Allow proper drying/curing time"]
            }
        ],
        "estimated_time": "2-3 hours",
        "difficulty": "intermediate",
        "materials_needed": product.get('materials', '').split(',') if product.get('materials') else []
    }