ancient_artisans/requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
Werkzeug==2.3.7
mysql-connector-python==8.0.33
google-generativeai==0.3.2
python-dotenv==1.0.0
cryptography==41.0.4